pyproject.toml
[tool.poetry]
name = "epic_events_crm"
version = "0.1.0"
description = "Simple CRM for learning purposes"
authors = ["epic_events"]
packages = [
	{ include = "app" },
	{ include = "cli" }
]

[tool.poetry.dependencies]
python = "^3.13"
sqlalchemy = "*"
pymysql = "*"
python-dotenv = "*"
argon2-cffi = "*"
pyjwt = "*"
click = "*"
cryptography = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
